% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/null.logplatt.gev.fit.R
\name{null.logplatt.gevd.fit}
\alias{null.logplatt.gevd.fit}
\title{null.logplatt.gevd.fit}
\usage{
null.logplatt.gevd.fit(platt.null.vec, standardizeQ = FALSE, plotQ = FALSE)
}
\arguments{
\item{platt.null.vec}{Vector of null (non-match) Platt scores from an SVM}

\item{standardizeQ}{Whether or not to standardize the log null scores}

\item{plotQ}{Diagnostic plots?}
}
\value{
A list with the fitted parameters, fit info and chi-square goodness of fit test results
}
\description{
Generalized extreme value fit for log null Platt scores.
}
\details{
We noticed that the log of the null (non-match) Platt scores typically looks close to Gaussian.
However we've also noticed that this distribution can show some skew. The generalized extreme value
distribution is a flexible three parameter distribution which usually gives a good fit to the null log
Platt score distribution estimated by the bootstrapping procedure. This routine calls the \code{fgev} function
from the evd package.

It is not necessary to fit the log of the null (non-match) Platt scores as p-values with respect to them can
be obtained empirically (cf. Storey and Tibshirani). Also other fits are possible. It is up to the user if
they choose to use this function to help obtain p-values.

In practice, we use this function because (1) it avoids p-values = 0 (always encountered when obtaining empirical p-values
for non-null (known-matching) scores), (2) it is faster than computing empirical p-values from hundres of
thousands or millions of bootstrapped null scores and (3) it cuts down on the spagetti code for the
distribution's fit when we do it step-by-step.
}
\examples{
XXXX
}
\references{
Storey JD, Tibshirani R. Statistical significance for genomewide studies. PNAS 100(16):9440-9445 (2003)
}

