% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/null.logplatt.nig.fit.R
\name{null.logplatt.nig.fit}
\alias{null.logplatt.nig.fit}
\title{null.logplatt.nig.fit}
\usage{
null.logplatt.nig.fit(platt.null.vec, alpha.init = NULL, beta.init = NULL,
  delta.init = NULL, mu.init = NULL, standardizeQ = FALSE,
  plotQ = FALSE)
}
\arguments{
\item{platt.null.vec}{Vector of null (non-match) Platt scores from an SVM}

\item{alpha.init}{Initial guess of alpha parameter for NIG}

\item{beta.init}{Initial guess of beta parameter for NIG}

\item{delta.init}{Initial guess of delta parameter for NIG}

\item{mu.init}{Initial guess of mu parameter for NIG}

\item{standardizeQ}{Whether or not to standardize the log null scores}

\item{plotQ}{Diagnostic plots?}
}
\value{
A list with the fitted parameters, fit info and chi-square goodness of fit test results
}
\description{
Normal inverse Gaussian fit for log null Platt scores.
}
\details{
We noticed that the log of the null (non-match) Platt scores typically looks close to Gaussian.
However we've also noticed that this distribution can show some skew. The normal inverse Gaussian (nig)
distribution is a flexible four parameter distribution which usually gives a good fit to the null log
Platt score distribution estimated by the bootstrapping procedure. This routine calls the \code{nigFit} function
in "mle" mode from the fBasics package.

It is not necessary to fit the log of the null (non-match) Platt scores as p-values with respect to them can
be obtained empirically (cf. Storey and Tibshirani). Also other fits are possible. It is up to the user if
they choose to use this function to help obtain p-values.

In practice, we use this function because (1) it avoids p-values = 0 (always encountered when obtaining empirical p-values
for non-null (known-matching) scores), (2) it is faster than computing empirical p-values from hundres of
thousands or millions of bootstrapped null scores and (3) it cuts down on the spagetti code for the
distribution's fit when we do it step-by-step.
}
\examples{
XXXX
}
\references{
Storey JD, Tibshirani R. Statistical significance for genomewide studies. PNAS 100(16):9440-9445 (2003)
}

