% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/null.logplatt.sn.fit.R
\name{null.logplatt.sn.fit}
\alias{null.logplatt.sn.fit}
\title{null.logplatt.sn.fit}
\usage{
null.logplatt.sn.fit(platt.null.vec, standardizeQ = FALSE, plotQ = FALSE)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
Skewed normal fit for log null Platt scores.
}
\details{
XXXX
}
\examples{
XXXX
}
\references{
Storey JD, Tibshirani R. Statistical significance for genomewide studies. PNAS 100(16):9440-9445 (2003)
}

