% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/null.logplatt.gaussian.fit.R
\name{null.platt.gaussian.fit}
\alias{null.platt.gaussian.fit}
\title{Fit log of null Platt scores to a Gaussian}
\usage{
null.platt.gaussian.fit(platt.null.vec, standardizeQ = FALSE, plotQ = FALSE)
}
\arguments{
\item{platt.null.vec}{Vector of null (non-match) Platt scores from an SVM}

\item{standardizeQ}{Whether or not to standardize the log null scores}

\item{plotQ}{Diagnostic plots?}
}
\value{
A list with the fitted parameters, fit info and chi-square goodness of fit test results
}
\description{
Gaussian fit for log null Platt scores.
}
\details{
Simple wrapper to get the Gaussian fit around the log of the null Platt scores.
}
\examples{
XXXX
}
\references{
Storey JD, Tibshirani R. Statistical significance for genomewide studies. PNAS 100(16):9440-9445 (2003)
}

