% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/posterior.probs.for.unks2.R
\name{posterior.probs.for.unks2}
\alias{posterior.probs.for.unks2}
\title{XXXX}
\usage{
posterior.probs.for.unks2(training.dmat, training.labels, C.param, test.dmat,
  standardizeQ = FALSE, mean.log.null.score = NULL,
  sd.log.null.score = NULL, pvalue.method = "empirical",
  null.vec.training = NULL, distribution.name = NULL,
  distribution.fit.info = NULL, pp.point.est.func, pp.uci.est.func,
  pp.lci.est.func)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXXX
}
\details{
XXXX
}
\examples{
XXXX
}
\references{
Storey JD, Tibshirani R. Statistical significance for genomewide studies. PNAS 100(16):9440-9445 (2003)
}

