% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zscore.fit.prep.R
\name{zscore.fit.prep}
\alias{zscore.fit.prep}
\title{XXXX}
\usage{
zscore.fit.prep(precomputed.null.scores = NULL, training.dmat,
  validation.dmat, training.labels, validation.labels,
  distribution = "gaussian", num.processes, standardizeQ = TRUE,
  num.bs.iter = 2000, C.param = 0.1, printQ = FALSE, plotQ = FALSE)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXXX
}
\details{
XXXX
}
\examples{
XXXX
}
\references{
Storey JD, Tibshirani R. Statistical significance for genomewide studies. PNAS 100(16):9440-9445 (2003)
}

