% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zscore.fit.R
\name{zscore.fit}
\alias{zscore.fit}
\title{Construct z-scores using a parametric fit distribution to the null Platt scores}
\usage{
zscore.fit(precomputed.null.scores = NULL, training.dmat, validation.dmat,
  training.labels, validation.labels, distribution = "gaussian",
  num.processes, standardizeQ = TRUE, num.bs.iter = 2000, C.param = 0.1,
  printQ = FALSE, plotQ = FALSE)
}
\arguments{
\item{XXXX}{XXXXXXX}
}
\value{
XXXX
}
\description{
Construct z-scores using a parametric fit distribution to the null Platt scores
}
\details{
A wrapper for the procedure to get null and non-null z-scores from a parametric fit around the bootstrapped
null (log) Platt scores. The log of the boostrapped null scores will be taken automatically.
}
\examples{
XXXX
}
\references{
Storey JD, Tibshirani R. Statistical significance for genomewide studies. PNAS 100(16):9440-9445 (2003)
}

