% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zscore.fit2.R
\name{zscore.fit2}
\alias{zscore.fit2}
\title{Construct z-scores using a parametric fit distribution to the null Platt scores}
\usage{
zscore.fit2(precomputed.null.scores = NULL, training.dmat, validation.dmat,
  training.labels, validation.labels, distribution = "gaussian",
  num.processes, standardizeQ = TRUE, num.bs.iter = 2000, C.param = 0.1,
  pvalue.method = "empirical", printQ = FALSE, plotQ = FALSE)
}
\arguments{
\item{precomputed.null.scores}{Optional vector of precomputed Null Platt scores.}

\item{training.dmat}{Training set feature vector matrix}

\item{validation.dmat}{Validation set feature vector matrix}

\item{training.labels}{Training set labels vector}

\item{validation.labels}{Validation set labels vector}

\item{num.processes}{Number of processes to use if computing Null Platt scores to use in the z-score generation process.}

\item{standardizeQ}{Standardize the Log Platt scores wrt training set log Null Platt scores?}

\item{num.bs.iter}{Number of iterations to use if computing Null Platt scores to use in the z-score generation process.}

\item{C.param}{C-Penalty parameter for the linear kernel SVM steps.}

\item{pvalue.method}{Method to compute requisite p-values. Choices are "empirical" or "integral". See details below for warnings.}

\item{printQ}{Print out intermediate values/diagnostic info?}

\item{plotQ}{Display diagnostic plots after each major step in the z-score generation process?}

\item{distribution="lg"}{Parametric distribution to fit training set log Null Platt scores to. Choices are: "gev" (generalized.extreme.value), "nig" (normal.inverse.gaussian), "sn" (skew.normal) or "lg" (gaussian)}
}
\value{
all.score.calc.info      A list containing: standardization.flag, fit.distribution.name, fit.info.and.diagnostics,
boostrapped.null.platt.scores, score.null.training, score.null.validation, score.nonnull.validation, null.pvalues,
nonnull.pvalues (UN-SMEARED!!!!!!), null.zvalues, nonnull.zvalues.smeared (SMEARED!!!!!!)
}
\description{
Construct z-scores using a parametric fit distribution to the null Platt scores
}
\details{
A wrapper for the procedure to get null and non-null z-scores from a parametric fit around the bootstrapped
null (log) Platt scores. The log of the boostrapped null scores will be taken automatically.
}
\examples{
XXXX
}
\references{
Storey JD, Tibshirani R. Statistical significance for genomewide studies. PNAS 100(16):9440-9445 (2003)
}

