% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zscore.unk.R
\name{zscore.unk}
\alias{zscore.unk}
\title{XXXX}
\usage{
zscore.unk(test.score.vec, pvalue.method, null.vec.training = NULL,
  distribution = NULL, dist.fit.info = NULL)
}
\arguments{
\item{XXXX}{}
}
\value{
XXXX
}
\description{
XXXXX
}
\details{
XXXX
}
\examples{
XXXX
}
\references{
Storey JD, Tibshirani R. Statistical significance for genomewide studies. PNAS 100(16):9440-9445 (2003)
}

